Move plot methods to GLM/GEE classes from base
